Rename get invoice line
